:recycle: Move logger to index.ts
