+++ src/pipes/get-latest-version.ts
@@ -1,6 +1,5 @@
 import type { IEither } from '../utils/either'
-import type { ILogger, IColorizer } from '../utils/logger'
-import type { Unary } from '../types/common-types'
+import type { Unary, ILogger, IColorizer } from '../types/common-types'
 import type { IAppCtx } from 'types/app-ctx'
 import { tap } from '../utils/helpers'
